+++ src/main.js
@@ -4,19 +4,16 @@ import {numberWatched} from './components/data-profile.js';
 import {Navigation} from './components/navigation.js';
 import {arrayNavigation} from './components/data-navigation.js';
 import {getElementSort} from './components/sort.js';
-import {Film} from './components/film-card.js';
 import {getElementShowMore} from './components/button-show-more.js';
-import {FilmDetails} from './components/film-details.js';
-import {arrayComents} from './components/data-comments.js';
-import {Comments} from './components/film-details-comments.js';
 import {arrayFilms} from './components/array-films.js';
 import {totalCard} from './components/array-films.js';
-import {render, unrender} from './components/utils.js';
+import {render} from './components/utils.js';
+import {PageController} from './components/page.js';
 
 
 const header = document.querySelector(`.header`);
 const main = document.querySelector(`.main`);
-const sectionFilms = document.createElement(`section`).classList.add(`films`);
+const sectionFilms = document.createElement(`section`);
 const sectionFilmsList = document.createElement(`section`);
 const filmListTitle = `<h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>`;
 const filmListContainer = document.createElement(`div`);
@@ -26,7 +23,6 @@ const filmListContainerTopRated = document.createElement(`div`);
 const sectionFilmListMostCommented = document.createElement(`section`);
 const filmListMostCommentedTitle = `<h2 class="films-list__title">Most commented</h2>`;
 const filmListContainerMostCommented = document.createElement(`div`);
-const body = document.querySelector(`body`);
 const navigation = document.createElement(`nav`);
 const footerCount = document.querySelector(`.footer__statistics`);
 
@@ -39,7 +35,9 @@ const renderShowMore = () => {
     const number = 5;
     const pageCount = filmListContainer.querySelectorAll(`.film-card`).length;
 
-    arrayFilms.slice(pageCount, pageCount + number).forEach((filmMock) => renderFilmCard(filmMock, filmListContainer));
+    const pageController = new PageController(filmListContainer, arrayFilms.slice(pageCount, pageCount + number));
+    pageController.init();
+
     if (pageCount >= totalCard) {
       load.classList.add(`visually-hidden`);
     }
@@ -65,76 +63,6 @@ const renderNavigation = (navMock) => {
   render(navigation, nav.getElement());
 };
 
-const renderFilmCard = (filmMock, where) => {
-  const film = new Film(filmMock);
-  const filmDetails = new FilmDetails(filmMock);
-
-  if (!sectionFilmsList.querySelector(`.films-list__container`)) {
-    filmListContainer.classList.add(`films-list__container`);
-    sectionFilmsList.appendChild(filmListContainer);
-  }
-
-  let commentsList = null;
-
-  const renderComment = (commentMock) => {
-    const comment = new Comments(commentMock);
-    commentsList = document.querySelector(`.film-details__comments-list`);
-    render(commentsList, comment.getElement());
-  };
-
-  const onEscKeyDown = (evt) => {
-    if (evt.key === `Escape` || evt.key === `Esc`) {
-      unrender(filmDetails.getElement());
-      commentsList.innerHTML = ``;
-      document.removeEventListener(`keydown`, onEscKeyDown);
-    }
-  };
-
-  film.getElement()
-    .querySelector(`img`)
-    .addEventListener(`click`, () => {
-      render(body, filmDetails.getElement());
-      arrayComents.forEach((commentMock) => renderComment(commentMock));
-      document.addEventListener(`keydown`, onEscKeyDown);
-    });
-
-  film.getElement()
-    .querySelector(`.film-card__title`)
-    .addEventListener(`click`, () => {
-      render(body, filmDetails.getElement());
-      arrayComents.forEach((commentMock) => renderComment(commentMock));
-      document.addEventListener(`keydown`, onEscKeyDown);
-    });
-
-  film.getElement()
-    .querySelector(`.film-card__comments`)
-    .addEventListener(`click`, () => {
-      render(body, filmDetails.getElement());
-      arrayComents.forEach((commentMock) => renderComment(commentMock));
-      document.addEventListener(`keydown`, onEscKeyDown);
-    });
-
-  filmDetails.getElement().querySelector(`textarea`)
-    .addEventListener(`focus`, () => {
-      document.removeEventListener(`keydown`, onEscKeyDown);
-    });
-
-  filmDetails.getElement().querySelector(`textarea`)
-    .addEventListener(`blur`, () => {
-      document.addEventListener(`keydown`, onEscKeyDown);
-    });
-
-  filmDetails.getElement()
-    .querySelector(`.film-details__close-btn`)
-    .addEventListener(`click`, () => {
-      unrender(filmDetails.getElement());
-      commentsList.innerHTML = ``;
-      document.removeEventListener(`keydown`, onEscKeyDown);
-    });
-
-  render(where, film.getElement());
-};
-
 const renderFilmsExtra = (section, container, title, first, second) => {
   section.classList.add(`films-list--extra`);
   sectionFilms.appendChild(section);
@@ -144,7 +72,8 @@ const renderFilmsExtra = (section, container, title, first, second) => {
   container.classList.add(`films-list__container`);
   section.appendChild(container);
 
-  arrayFilms.slice(first, second).forEach((filmMock) => renderFilmCard(filmMock, container));
+  const pageControllerExtra = new PageController(container, arrayFilms.slice(first, second));
+  pageControllerExtra.init();
 };
 
 addComponent(header, getElementSearch());
@@ -153,7 +82,7 @@ arrayNavigation.forEach((navMock) => renderNavigation(navMock));
 
 addComponent(main, getElementSort());
 
-//sectionFilms.classList.add(`films`);
+sectionFilms.classList.add(`films`);
 main.appendChild(sectionFilms);
 
 sectionFilmsList.classList.add(`films-list`);
@@ -161,7 +90,10 @@ sectionFilms.appendChild(sectionFilmsList);
 
 addComponent(sectionFilmsList, filmListTitle);
 
-arrayFilms.slice(0, 5).forEach((filmMock) => renderFilmCard(filmMock, filmListContainer));
+filmListContainer.classList.add(`films-list__container`);
+sectionFilmsList.appendChild(filmListContainer);
+const pageController = new PageController(filmListContainer, arrayFilms.slice(0, 5));
+pageController.init();
 
 renderFilmsExtra(sectionFilmListTopRated, filmListContainerTopRated, filmListTopRatedTitle, 0, 2);
 renderFilmsExtra(sectionFilmListMostCommented, filmListContainerMostCommented, filmListMostCommentedTitle, 3, 5);
